backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
pydantic==2.11.7
psycopg2-binary==2.9.10
python-multipart==0.0.20
openai-whisper==20250625
python-dotenv==1.1.1
celery
redis==6.4.0
